requirements.txt
flask==3.0.3
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
numpy==2.1.0
